# openapi.yaml
openapi: 3.0.0
info:
  title: Sistema de Análise de Estatísticas de Futebol
  description: |
    API para análise detalhada de estatísticas de futebol com integração
    aos dados da VStats API. Fornece comparativas de desempenho home/away
    e métricas de estabilidade (Coeficiente de Variação).
  version: 1.2.0
  contact:
    name: Suporte
    email: contato@palpitremestre.com
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Ambiente de desenvolvimento
  - url: https://api.palpitremestre.com
    description: Produção
paths:
  /api/partidas:
    get:
      summary: Lista partidas por data
      description: |
        Retorna todas as partidas agendadas para uma data específica
        de todas as competições suportadas.
      operationId: listar_partidas
      tags:
        - Partidas
      parameters:
        - name: data
          in: query
          required: true
          description: Data no formato YYYY-MM-DD
          schema:
            type: string
            format: date
            example: "2025-12-27"
      responses:
        '200':
          description: Lista de partidas retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartidaListResponse'
        '400':
          description: Data inválida ou formato incorreto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro ao buscar da API VStats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-code-samples:
        - lang: javascript
          source: |
            const response = await fetch('http://localhost:8000/api/partidas?data=2025-12-27');
            const data = await response.json();
  /api/partida/{matchId}/stats:
    get:
      summary: Estatísticas detalhadas de uma partida
      description: |
        Retorna estatísticas agregadas (médias e CVs) de um time em um período
        específico (geral, últimas 5 ou últimas 10 partidas).
      operationId: get_stats
      tags:
        - Estatísticas
      parameters:
        - name: matchId
          in: path
          required: true
          description: ID único da partida
          schema:
            type: string
            example: f4vscquffy37afgv0arwcbztg
        - name: filtro
          in: query
          required: false
          description: |
            Período de análise (quantidade máxima de jogos disputados por time).

            - geral: até 50 jogos
            - 5: até 5 jogos
            - 10: até 10 jogos

            Observação: apenas partidas com data válida e placar disponível entram na amostra. Se houver menos jogos do que o limite, o cálculo usa os jogos disponíveis.

            Fallback: se não houver partidas individuais suficientes (ou a VStats falhar ao retornar stats por partida), o backend usa agregados da temporada (seasonstats) e registra `seasonstats_fallback` em `contexto.ajustes_aplicados`. Nesse modo, as contagens de amostra podem refletir a temporada (e podem ser > 5/10).
          schema:
            type: string
            enum:
              - geral
              - "5"
              - "10"
            default: geral
        - name: periodo
          in: query
          required: false
          description: |
            Tempo do jogo para análise.

            Observação: se o provedor não fornecer estatísticas por tempo (1T/2T) para uma partida, o backend faz fallback para o total do jogo (integral) e registra o ajuste em `contexto.ajustes_aplicados` (ex.: `periodo_fallback_integral`).

            Observação: se o backend estiver em fallback de seasonstats (`seasonstats_fallback`), o recorte por período pode não ser aplicável (a base vira agregado).
          schema:
            type: string
            enum:
              - integral
              - 1T
              - 2T
            default: integral
        - name: home_mando
          in: query
          required: false
          description: Subfiltro de mando para o mandante (segmenta a amostra por casa/fora)
          schema:
            type: string
            enum:
              - casa
              - fora
            nullable: true
        - name: away_mando
          in: query
          required: false
          description: Subfiltro de mando para o visitante (segmenta a amostra por casa/fora)
          schema:
            type: string
            enum:
              - casa
              - fora
            nullable: true
      responses:
        '200':
          description: Estatísticas calculadas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
        '400':
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Partida não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro ao calcular estatísticas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/partida/{matchId}/analysis:
    get:
      summary: Analise consolidada (stats + previsoes + over/under)
      description: |
        Retorna as estatisticas agregadas (StatsResponse) junto do modelo de previsao
        (Min/Exata/Max) e das probabilidades Over/Under para mercados de aposta.
      operationId: get_analysis
      tags:
        - Estatísticas
      parameters:
        - name: matchId
          in: path
          required: true
          description: ID único da partida
          schema:
            type: string
            example: f4vscquffy37afgv0arwcbztg
        - name: filtro
          in: query
          required: false
          description: |
            Período de análise (quantidade máxima de jogos disputados por time).

            - geral: até 50 jogos
            - 5: até 5 jogos
            - 10: até 10 jogos

            Observação: apenas partidas com data válida e placar disponível entram na amostra. Se houver menos jogos do que o limite, o cálculo usa os jogos disponíveis.

            Fallback: se não houver partidas individuais suficientes (ou a VStats falhar ao retornar stats por partida), o backend usa agregados da temporada (seasonstats) e registra `seasonstats_fallback` em `contexto.ajustes_aplicados`. Nesse modo, as contagens de amostra podem refletir a temporada (e podem ser > 5/10).
          schema:
            type: string
            enum:
              - geral
              - "5"
              - "10"
            default: geral
        - name: periodo
          in: query
          required: false
          description: |
            Tempo do jogo para análise.

            Observação: se o provedor não fornecer estatísticas por tempo (1T/2T) para uma partida, o backend faz fallback para o total do jogo (integral) e registra o ajuste em `contexto.ajustes_aplicados` (ex.: `periodo_fallback_integral`).

            Observação: se o backend estiver em fallback de seasonstats (`seasonstats_fallback`), o recorte por período pode não ser aplicável (a base vira agregado).
          schema:
            type: string
            enum:
              - integral
              - 1T
              - 2T
            default: integral
        - name: home_mando
          in: query
          required: false
          description: Subfiltro de mando para o mandante (segmenta a amostra por casa/fora)
          schema:
            type: string
            enum:
              - casa
              - fora
            nullable: true
        - name: away_mando
          in: query
          required: false
          description: Subfiltro de mando para o visitante (segmenta a amostra por casa/fora)
          schema:
            type: string
            enum:
              - casa
              - fora
            nullable: true
        - name: debug
          in: query
          required: false
          description: |
            Quando `1` (ou `true`), inclui `debug_amostra` na resposta com IDs/datas/pesos das partidas usadas no cálculo do recorte (útil para auditoria e envio para IA).
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Analise calculada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsAnalysisResponse'
        '400':
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Partida não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro ao calcular analise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/time/{teamId}/escudo:
    get:
      summary: Escudo/logo de um time
      description: Retorna URL do escudo e nome do time
      operationId: get_escudo
      tags:
        - Times
      parameters:
        - name: teamId
          in: path
          required: true
          description: ID do time
          schema:
            type: string
            example: 4dsgumo7d4zupm2ugsvm4zm4d
        - name: nome
          in: query
          required: false
          description: Nome do time (fallback)
          schema:
            type: string
            example: Arsenal
      responses:
        '200':
          description: Escudo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EscudoResponse'
        '404':
          description: Escudo não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro ao buscar escudo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    TimeInfo:
      type: object
      required:
        - id
        - nome
        - codigo
      properties:
        id:
          type: string
          description: ID único do time
          example: 4dsgumo7d4zupm2ugsvm4zm4d
        nome:
          type: string
          description: Nome do time
          example: Arsenal
        codigo:
          type: string
          description: Código do time
          example: ARS
        escudo:
          type: string
          format: uri
          nullable: true
          description: URL do escudo
          example: https://r2.thesportsdb.com/images/media/team/badge/uyhbfe1612467038.png
    PartidaResumo:
      type: object
      required:
        - id
        - data
        - horario
        - competicao
        - estadio
        - mandante
        - visitante
      properties:
        id:
          type: string
          example: f4vscquffy37afgv0arwcbztg
        data:
          type: string
          format: date
          example: "2025-12-27"
        horario:
          type: string
          format: time
          example: "17:00"
        competicao:
          type: string
          example: Premier League
        estadio:
          type: string
          example: Emirates Stadium
        mandante:
          $ref: '#/components/schemas/TimeInfo'
        visitante:
          $ref: '#/components/schemas/TimeInfo'
    PartidaListResponse:
      type: object
      required:
        - data
        - total_partidas
        - partidas
      properties:
        data:
          type: string
          format: date
          example: "2025-12-27"
        total_partidas:
          type: integer
          minimum: 0
          example: 2
        partidas:
          type: array
          items:
            $ref: '#/components/schemas/PartidaResumo'
    EstatisticaMetrica:
      type: object
      required:
        - media
        - cv
        - classificacao
        - estabilidade
      properties:
        media:
          type: number
          format: float
          minimum: 0
          example: 5.88
        cv:
          type: number
          format: float
          minimum: 0
          example: 0.32
        classificacao:
          type: string
          enum:
            - Muito Estável
            - Estável
            - Moderado
            - Instável
            - Muito Instável
            - N/A
          example: Moderado
        estabilidade:
          type: integer
          minimum: 0
          maximum: 100
          example: 68
    EstatisticaFeitos:
      type: object
      required:
        - feitos
        - sofridos
      properties:
        feitos:
          $ref: '#/components/schemas/EstatisticaMetrica'
        sofridos:
          $ref: '#/components/schemas/EstatisticaMetrica'
    EstatisticasTime:
      type: object
      required:
        - escanteios
        - gols
        - finalizacoes
        - finalizacoes_gol
        - cartoes_amarelos
        - faltas
      properties:
        escanteios:
          $ref: '#/components/schemas/EstatisticaFeitos'
        gols:
          $ref: '#/components/schemas/EstatisticaFeitos'
        finalizacoes:
          $ref: '#/components/schemas/EstatisticaFeitos'
        finalizacoes_gol:
          $ref: '#/components/schemas/EstatisticaFeitos'
        cartoes_amarelos:
          $ref: '#/components/schemas/EstatisticaMetrica'
        faltas:
          $ref: '#/components/schemas/EstatisticaMetrica'
    TimeComEstatisticas:
      type: object
      required: [id, nome, estatisticas, recent_form]
      properties:
        id:
          type: string
        nome:
          type: string
        escudo:
          type: string
          nullable: true
        estatisticas:
          $ref: '#/components/schemas/EstatisticasTime'
        recent_form:
          type: array
          description: Sequência de resultados recentes (W=win, D=draw, L=loss)
          items:
            type: string
            enum: [W, D, L]
    ArbitroInfo:
      type: object
      required: [id, nome, partidas, partidas_temporada, media_cartoes_amarelos, media_cartoes_temporada]
      properties:
        id:
          type: string
        nome:
          type: string
        partidas:
          type: integer
          minimum: 0
        partidas_temporada:
          type: integer
          minimum: 0
        media_cartoes_amarelos:
          type: number
          minimum: 0
        media_cartoes_temporada:
          type: number
          minimum: 0
        media_faltas:
          type: number
          minimum: 0
          nullable: true
    StatsResponse:
      type: object
      required:
        - partida
        - filtro_aplicado
        - partidas_analisadas
        - mandante
        - visitante
      properties:
        partida:
          $ref: '#/components/schemas/PartidaResumo'
        filtro_aplicado:
          type: string
          enum:
            - geral
            - "5"
            - "10"
          example: "5"
        partidas_analisadas:
          type: integer
          minimum: 1
          example: 5
        partidas_analisadas_mandante:
          type: integer
          minimum: 0
          nullable: true
          description: Numero de partidas usadas para o mandante (pode diferir por subfiltros). Quando `seasonstats_fallback` estiver ativo, pode refletir a temporada (agregado) e ser > 5/10.
        partidas_analisadas_visitante:
          type: integer
          minimum: 0
          nullable: true
          description: Numero de partidas usadas para o visitante (pode diferir por subfiltros). Quando `seasonstats_fallback` estiver ativo, pode refletir a temporada (agregado) e ser > 5/10.
        mandante:
          $ref: '#/components/schemas/TimeComEstatisticas'
        visitante:
          $ref: '#/components/schemas/TimeComEstatisticas'
        arbitro:
          $ref: '#/components/schemas/ArbitroInfo'
          nullable: true
        contexto:
          $ref: '#/components/schemas/ContextoPartida'
          nullable: true
        h2h_all_comps:
          $ref: '#/components/schemas/H2HInfo'
          nullable: true
        debug_amostra:
          $ref: '#/components/schemas/DebugAmostra'
          nullable: true
    DebugAmostraTime:
      type: object
      required:
        - source
        - limit
        - periodo
        - mando
        - n_used
        - match_ids
        - match_dates
        - weights
      properties:
        source:
          type: string
          enum: [matches, seasonstats_fallback]
        limit:
          type: integer
          minimum: 0
          description: Limite solicitado (5/10/50), antes de dedupe/falhas
        periodo:
          type: string
          enum: [integral, 1T, 2T]
        mando:
          type: string
          enum: [casa, fora]
          nullable: true
        n_used:
          type: integer
          minimum: 0
          description: Tamanho efetivo da amostra usada no cálculo
        match_ids:
          type: array
          items:
            type: string
        match_dates:
          type: array
          items:
            type: string
            nullable: true
        weights:
          type: array
          items:
            type: number
        reason:
          type: string
          nullable: true
          description: Motivo do fallback/ajuste (opcional)
    DebugAmostra:
      type: object
      required: [mandante, visitante]
      properties:
        mandante:
          $ref: '#/components/schemas/DebugAmostraTime'
        visitante:
          $ref: '#/components/schemas/DebugAmostraTime'
    ContextoTime:
      type: object
      properties:
        dias_descanso:
          type: integer
          minimum: 0
          nullable: true
        jogos_7d:
          type: integer
          minimum: 0
          nullable: true
        jogos_14d:
          type: integer
          minimum: 0
          nullable: true
        posicao:
          type: integer
          minimum: 1
          nullable: true
    ClassificacaoTimeInfo:
      type: object
      required: [posicao]
      properties:
        posicao:
          type: integer
          minimum: 1
        pontos:
          type: integer
          minimum: 0
        jogos:
          type: integer
          minimum: 0
        saldo_gols:
          type: integer
          nullable: true
    H2HInfo:
      type: object
      required: [total_matches, avg_goals_per_match]
      properties:
        total_matches:
          type: integer
          minimum: 0
        avg_goals_per_match:
          type: number
          minimum: 0
        home_wins:
          type: integer
          minimum: 0
        away_wins:
          type: integer
          minimum: 0
        draws:
          type: integer
          minimum: 0
    ContextoPartida:
      type: object
      required: [mandante, visitante]
      properties:
        mandante:
          $ref: '#/components/schemas/ContextoTime'
        visitante:
          $ref: '#/components/schemas/ContextoTime'
        classificacao_mandante:
          $ref: '#/components/schemas/ClassificacaoTimeInfo'
        classificacao_visitante:
          $ref: '#/components/schemas/ClassificacaoTimeInfo'
        h2h_all_comps:
          $ref: '#/components/schemas/H2HInfo'
        fase:
          type: string
          nullable: true
        ajustes_aplicados:
          type: array
          description: |
            Lista de ajustes/fallbacks aplicados pelo backend para manter a resposta consistente.

            Valores conhecidos:
            - `periodo_fallback_integral`: recorte 1T/2T indisponível em parte da amostra; usamos integral.
            - `seasonstats_fallback`: sem partidas individuais suficientes; usamos agregado de temporada (seasonstats).
          items:
            type: string
          example: ["periodo_fallback_integral"]
...
    PrevisaoEstatistica:
      type: object
      required: [home, away, total]
      properties:
        home:
          $ref: '#/components/schemas/PrevisaoValor'
        away:
          $ref: '#/components/schemas/PrevisaoValor'
        total:
          $ref: '#/components/schemas/PrevisaoValor'
    PrevisaoPartida:
      type: object
      required:
        - gols
        - escanteios
        - finalizacoes
        - finalizacoes_gol
        - cartoes_amarelos
        - faltas
      properties:
        gols:
          $ref: '#/components/schemas/PrevisaoEstatistica'
        escanteios:
          $ref: '#/components/schemas/PrevisaoEstatistica'
        finalizacoes:
          $ref: '#/components/schemas/PrevisaoEstatistica'
        finalizacoes_gol:
          $ref: '#/components/schemas/PrevisaoEstatistica'
        cartoes_amarelos:
          $ref: '#/components/schemas/PrevisaoEstatistica'
        faltas:
          $ref: '#/components/schemas/PrevisaoEstatistica'
...
    OverUnderPartida:
      type: object
      required:
        - gols
        - escanteios
        - finalizacoes
        - finalizacoes_gol
        - cartoes_amarelos
        - faltas
      properties:
        gols:
          $ref: '#/components/schemas/OverUnderStat'
        escanteios:
          $ref: '#/components/schemas/OverUnderStat'
        finalizacoes:
          $ref: '#/components/schemas/OverUnderStat'
        finalizacoes_gol:
          $ref: '#/components/schemas/OverUnderStat'
        cartoes_amarelos:
          $ref: '#/components/schemas/OverUnderStat'
        faltas:
          $ref: '#/components/schemas/OverUnderStat'
    StatsAnalysisResponse:
      allOf:
        - $ref: '#/components/schemas/StatsResponse'
        - type: object
          required: [previsoes, over_under]
          properties:
            previsoes:
              $ref: '#/components/schemas/PrevisaoPartida'
            over_under:
              $ref: '#/components/schemas/OverUnderPartida'
...
    EscudoResponse:
      type: object
      required:
        - team_id
        - team_name
        - escudo_url
        - fonte
      properties:
        team_id:
          type: string
          example: 4dsgumo7d4zupm2ugsvm4zm4d
        team_name:
          type: string
          example: Arsenal
        escudo_url:
          type: string
          format: uri
          nullable: true
          example: https://r2.thesportsdb.com/images/media/team/badge/uyhbfe1612467038.png
        fonte:
          type: string
          example: TheSportsDB
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Descrição do erro
          example: "Data inválida: 2025-13-01 (formato deve ser YYYY-MM-DD)"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Autenticação JWT (futuro)
...